requirements.txt
torch
transformers
accelerate
sentencepiece
numpy
sounddevice
webrtcvad
openai-whisper
pyttsx3
TTS
python-dotenv
firebase-admin
pandas
langchain-core
langchain-community
langchain-classic
twilio
flask
flask-sock
pyngrok
#audioop